latest deploy custom expand datatable

+++ app/Http/Controllers/KOP/Helpers/RumusHitungBIayaListrik.php
@@ -154,7 +154,7 @@ class RumusHitungBIayaListrik {
 
     public static function HitungPPJ($totalcosperbulan){
 
-        return ( ($totalcosperbulan) + ($totalcosperbulan*0.03) + 6000 );
+        return ((float) ($totalcosperbulan) + ($totalcosperbulan*0.03) + 6000 );
 
     }
